Separated `Platform` into multiple files for better organization
